gc: do not ignore local lsn when persisting consumers
Currently, we do not persist local rows of consumers' vclocks because
they are not needed anyway - we ignore them on wal cleanup. But since
we are going to introduce checkpoint consumers, we should stop ignoring
local rows because they are part of checkpoint signature.

NO_CHANGELOG=internal
NO_DOC=internal