lua: prepare utils for box error trace improvement
Here is just of bunch utility functions that used in the patch that make
several modules throw box.error with trace set to the caller place. That
patch is just a boring huge switch to box.error and setting proper level
on error creation. Factor out utility functions so that they and their
tests are not get lost.

Part of #9914

NO_CHANGELOG=internal
NO_DOC=internal