console: check error trace where specified
In scope of the #9914 issue we are setting error trace for API to the
caller frame. Let's leverage existing diff tests to check error trace.
Just check error trace when error is raised when evaluating expression
in console which is used in diff tests.

The check is done for test build only and only for Lua modules specified
in `tarantool._internal.trace_check_is_required`.

Part of #9914

NO_TEST=internal
NO_CHANGELOG=internal
NO_DOC=internal