box: add `space::is_sync` (C) and `space.state.is_sync` (Lua) fields
In the scope of #9723 we are going to enforce synchronous replication for
system spaces when the synchronous queue is claimed. This means that the
`space_opts::is_sync` (C) and `space.is_sync` (Lua) fields will stop
reflecting the state of synchronous replication for system spaces.

Logically, the current `is_sync` fields reflect a user-provided option
rather than the state. Let's add `space::is_sync` (C) field and a read-only
`space.state` subtable with an `is_sync` field (Lua) that will reflect the
state of synchronous replication for system spaces. For user spaces, this
field will always mirror the `is_sync` option set for the space.

This approach allows flexibility in making system spaces synchronous on a
case-by-case basis. It will also allow to add some other state information
about spaces in the future. For instance, whether a space is currently
being altered.

Needed for #9723

NO_CHANGELOG=<in later commit>
NO_DOC=<in later commit>