relay: add option with_gc_consumer for subscribe
Now we update consumers only when sending data to non-anonymous
replicas. Since we are going to update gc consumer for anonymous
replicas that requested it with a special option in SUBSCRIBE,
let's add a new relay option `with_gc_consumer` - it controls
whether consumer should be updated on subscribe or not. Now it
is set for non-anonymous replicas and only for them but the next
commits will change that.

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring