memtx: refactor the tree index lookup out
The tree lookup consists of few steps: calling a BPS tree function
depending on the given key and iterator type and optional step back.
Prior to this patch, these actions were performed in the
`memtx_tree_index_iterator_start` function.

But it will also be required when the accelerated index count will
be implemented in the memtx tree, so it needs to be refactored out
for the sake of reusability.

This patch introduces a new `memtx_tree_lookup` function performing
the actions required for the primary lookup (this does not include
the step back for reverse iterators). The function is now used in
the `memtx_tree_index_iterator_start` instead.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring