memtx: refactor the tree index lookup out
Prior to this patch, the tree lookup had been performed in the
`memtx_tree_index_iterator_start` function, but it will also be
required when the accelerated index count will be implemented in
the memtx tree, so it needs to be refactored out for the sake of
reusability.

This patch refactors the key lookup logic into a new function:
`memtx_tree_lookup`, and makes the `memtx_tree_index_iterator_start`
use it.

Needed for #8204

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring