memtx: refactor the tree index lookup canonicalization out
During the iterator creation the memtx tree canonicalizes the given
key and iterator type in order to verify the input and restrict the
iterator type values. This includes the iterator type and key parts
check, type transforming and key canonicalization.

These steps are required for the fast index count to be implemented
in the memtx tree and for fixing the #10140, so it's refactored out
for the sake of reusability.

This patch introduces the `memtx_tree_lookup_canonicalize` function
which performs these transformations and validations and uses it in
the `memtx_tree_index_create_iterator`.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring