box: refactor system space identifiers definition
In the scope of #9723 we are going to need to iterate over a list of system
spaces. Let's refactor the system space identifiers definition into a
separate X macro to have one list to maintain. We will further add an
`is_sync` flag to each system space definition to generate a list of system
spaces for which synchronous replication should be enabled when the
synchronous queue is claimed.

Needed for #9723

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>