memtx: introduce the tuple count tracking
Currently performing count of tuples in an index (with or without
key) is not recorded in the memtx TX manager, which makes calling
`index:len` or `index:count` in a transaction with MVCC enabled
unsafe.

This patch introduces a new entity in the memtx transaction manager
to track such operations: GAP_COUNT, and a dedicated function for
creating them: memtx_tx_track_count.

The entity (gap item) is a record that a concurrent transaction has
counted tuples matching some key and iterator type in an index.

If a transaction creates such an entity, any insertion or deletion
of a matching tuple in the index will be conflicted with it. This
works differently for inserts and deletes:

1. If a concurrent transaction inserts a new matching tuple, then
   its read_gaps list is modified like the counted transaction had
   read the exact key of the tuple and found nothing. This creates
   a conflict.

2. If a concurrent transaction deletes a matching tuple, then the
   transaction that counted the tuple is inserted into the tuple
   reader list: it pretends to have read the tuple prior to the
   deletion. This creates a conflict.

NO_DOC=internal
NO_TEST=see the next commit
NO_CHANGELOG=internal