gc: link all consumers into a list
Currently, we index all active consumers by a tree. When Tarantool is
shutting down, we delete all consumers from that tree, so if a consumer
is inactive, it won't be deleted. So let's additionally link all
consumers into a list - all the consumers will be freed then. Also, we
will need this list when implementing persistent gc consumers.