gc: persist WAL GC consumers
The commit makes WAL GC consumers persistent. When the new consumer is
registered, it is not persistent - it will become one after the first
successful call of `gc_consumer_sync`. After the consumer has become
persistent, all its updates become asynchronous - persistent WAL GC
state is updated eventually by a background fiber.

Note that the new system space is treated in a special way - we don't
set our trigger for system spaces to it because it shouldn't acquire ddl
and limit depth of on_replace triggers. If we acquired ddl on replace in
_gc_consumers, user's transactions would be aborted for no apparent
reason when the consumers are advanced (that's what "acquire ddl"
actually do). And, since replace in `_gc_consumers` can be done from
on_replace trigger of `_cluster`, limiting depth of its triggers would
lead to error in such cases.

Since there is no need to wait limbo to update the local space, let's
use flag TXN_FORCE_ASYNC when possible. Moreover, we must be able to
subscribe a replica when quorum cannot be collected, so we must use
it to avoid deadlock when transaction synchronizing consumer on subscribe
waits for collected quorum but quorum cannot be collected until this
subscribe is done.

Function `gc_consumer_erase` is called only from on_replace trigger of
`_cluster`, hence, inside active transaction, so it never use this flag
since we mustn't set this flag to synchronous transactions.

Note that the commit completely breaks autorebootstrap - it does not
work anymore because local lsn is always bumped by persistent GC.
Corresponding test is dropped.

Closes #10154

@TarantoolBot document
Title: Persistent WAL GC state

Now all replicas has WAL GC consumers persisted in space `_gc_consumers`.
When the replica is deleted from space `_cluster`, its consumer is deleted
as well. It ensures that xlogs needed for replica won't be deleted after
restart of instance. Note that they still can be deleted when we have no
space for newer xlogs.

Space is read only on recovery, so inserting a new tuple to this space
won't have an effect until restart. If the space contains an invalid
tuple, it is skipped on recovery and an error is logged.