box: drop the `replicaset::is_joining` field as unused
The `replicaset::is_joining` field was previously used to track the
situation when a node joining a replicaset could get deleted from the
`_cluster` space. However, now in the scope of #10088 we always apply the
deletion of the replica from the `_cluster` space on the deleted replica,
so this field has become redundant.

Follows up #10088

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>