bps: deduplicate insertion functions
Prior to this patch there was three separated implementations of
insertion of a value in the tree: `insert`, `insert_get_iterator`
and `insert_get_offset`. This is error-prone and makes the further
development more difficult.

This patch introduces a new `bps_tree_insert_impl` function, which
deduplicates the existed insertion implementations and makes them
use it.

Now the value of `offset` parameter is undefined on error. This does
not affect the only user of the function, which is the memcs engine.

Part of #9982

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring