bps: deduplicate insertion functions
Prior to this patch there was three separated implementations of
insertion of a value in the tree: `insert`, `insert_get_iterator`
and `insert_get_offset`. This is error-prone and makes the further
development more difficult.

This patch introduces a new `bps_tree_insert_impl` function, which
deduplicates the existed insertion implementations and makes them
use it. The function is made `ALWAYS_INLINE` in order to prevent
regressions in the future.

Now the value of `offset` parameter is undefined on error. This does
not affect the only user of the function, which is the memcs engine.

Performance impact:

  | request | stat | old       | new       | diff   | stdev |
  | ------- | ---- | --------- | --------- | ------ | ----- |
  | insert  | RPS  | 953798.71 | 961295.51 | +0.79% | 0.33% |
  | select  | RPS  | 805361.94 | 807647.13 | +0.28% | 0.32% |
  | replace | RPS  | 953764.82 | 960956.13 | +0.75% | 0.33% |
  | delete  | RPS  | 833315.10 | 836079.60 | +0.33% | 0.35% |

Part of #9982

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring