bps: deduplicate deletion functions
Prior to this patch there was three separated implementations of
deletion of an element from the tree: `delete`, `delete_get_offset`
and `delete_value`. This is error-prone and makes the development
more difficult.

This patch introduces a new `bps_tree_delete_impl` function, which
deduplicates the existed implementations, and makes them use it. The
function is made `ALWAYS_INLINE` in order to prevent regressions.

By the way fixed the misleading comment for the `bps_tree_delete_value`
function and updated the `bps_tree_delete_get_offset` documentation: now
the value of the `offset` parameter is undefined on failure. This does
not affect the function behavior though.

Performance impact against the version prior to all refactorings:

  | request | stat | old       | new       | diff   | stdev |
  | ------- | ---- | --------- | --------- | ------ | ----- |
  | insert  | RPS  | 953798.71 | 964738.50 | +1.15% | 0.63% |
  | select  | RPS  | 805361.94 | 807050.13 | +0.21% | 0.66% |
  | replace | RPS  | 953764.82 | 964439.47 | +1.12% | 0.58% |
  | delete  | RPS  | 833315.10 | 834984.89 | +0.20% | 0.68% |

Part of #9982

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring