bps: deduplicate deletion functions
Prior to this patch there was three separated implementations of
deletion of an element from the tree: `delete`, `delete_get_offset`
and `delete_value`. This is error-prone and makes the development
more difficult.

This patch introduces a new `bps_tree_delete_impl` function, which
deduplicates the existed implementations, and makes them use it.

By the way fixed the misleading comment for the `bps_tree_delete_value`
function and updated the `bps_tree_delete_get_offset` documentation: now
the value of the `offset` parameter is undefined on failure. This does
not affect the function behavior though.

Part of #9982

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring