bps: deduplicate key lookup functions
Prior to this patch there was two separated implementations of the
key lookup sharing almost identical code: `bps_tree_find_impl` and
the `bps_tree_find_get_offset_impl`.

The only difference between the two functions is computing the offset
to the found element.

This patch removes the `bps_tree_find_get_offset_impl` and adds its
functionality to the `bps_tree_find_impl`.

By the way changed the interface so that the value of the `offset`
argument is undefined if the element is not found. This breaks the
BPS tree offset API tests, but this function is only used in the
memcs engine now, and it does not rely on the offset if no element
found.

Part of #9982

NO_DOC=refactoring
NO_CHANGELOG=refactoring