bps: introduce the bps_tree_max_blocks_required_for_inserts function
The function returns the max amount of new blocks that may be required
for insertion of N consequent elements into a tree. It may be used to
calculate the amount extents to reserve on batch insertion into the
tree.

The batch insertion is not implemetned yet, so this function calculates
the amount of blocks required for sequence of raw inserts.

NO_DOC=internal
NO_CHANGELOG=internal