bps: always touch the target leaf on insertion
On each insertion and deletion we go over all levels of the tree and
dereference block IDs using `bps_tree_restore_block`. The thing is,
in case of insertion, we may touch the target leaf unconditionally,
because there's no no-op flow in the `bps_tree_insert` after it:
we will either replace or insert an element into the leaf.

This patch makes the `bps_tree_collect_path` always touch the leaf
block in the `bps_tree_insert`. The function is now too complicated
for GCC 13 to inline it, so it's made `ALWAYS_INLINE` to prevent the
possible regressions.

Performance impact against the version prior to all commits:

  | request | stat | old       | new       | diff   | stdev |
  | ------- | ---- | --------- | --------- | ------ | ----- |
  | insert  | RPS  | 953798.71 | 957184.34 | +0.35% | 0.28% |
  | select  | RPS  | 805361.94 | 808730.63 | +0.42% | 0.23% |
  | replace | RPS  | 953764.82 | 957252.78 | +0.37% | 0.26% |
  | delete  | RPS  | 833315.10 | 835194.55 | +0.23% | 0.23% |

Part of #10322

NO_DOC=perf improvement
NO_TEST=perf improvement
NO_CHANGELOG=perf improvement