bps: don't lookup the matras ID if not required
We use the `matras_touch` function in order to notify the matras that
a block is going to be modified. In case the block has to be copied
in order to preserve read views, it allocates a new block and returns
its address, otherwise it just returns the current address of the
given block.

The thing is, if we have this address acquired already, we don't have
to look it up from the matras again sometimes.

Let's use the new matras API to check if an allocation of a new block
might be required and only update the pointer using the `matras_touch`
if it can be changed.

Performance impact against the version prior to all commits:

  | request | stat | old       | new       | diff   | stdev |
  | ------- | ---- | --------- | --------- | ------ | ----- |
  | insert  | RPS  | 953798.71 | 958990.66 | +0.54% | 0.47% |
  | select  | RPS  | 805361.94 | 808628.46 | +0.41% | 0.23% |
  | replace | RPS  | 953764.82 | 958759.03 | +0.52% | 0.43% |
  | delete  | RPS  | 833315.10 | 834287.96 | +0.12% | 0.23% |

Part of #10322

NO_DOC=perf improvement
NO_TEST=perf improvement
NO_CHANGELOG=perf improvement