bps: do not touch max_elem owners in the `bps_tree_touch_path`
In case if the insertion or deletion from the block requires the
reballancing or if the tree is used with a cardinality config, we
use the `bps_tree_touch_path` function in order to inform matras
about incoming block modifications.

The problem is that the function not only touches the leaf parents,
but also touches owners of `max_elem` data. But these owners are
included into the parents touched, so it's not required to touch
them again, we can simply use their new pointers after we have
touched all the path. This decreases the amount of matras ID
lookups in general (which is a pretty expensive operation).

Performance impact against the version prior to all commits:

  | request | stat | old       | new       | diff   | stdev |
  | ------- | ---- | --------- | --------- | ------ | ----- |
  | insert  | RPS  | 953798.71 | 978592.11 | +2.60% | 0.32% |
  | select  | RPS  | 805361.94 | 821532.64 | +2.01% | 0.23% |
  | replace | RPS  | 953764.82 | 978612.78 | +2.61% | 0.31% |
  | delete  | RPS  | 833315.10 | 842437.50 | +1.09% | 0.24% |

Part of #10322

NO_DOC=perf improvement
NO_TEST=perf improvement
NO_CHANGELOG=perf improvement