build: configure parallel jobs
It is not possible to get a number of parallel jobs in CMake
passed by the user with option "--parallel", but CMake allows to
pass a number of parallel jobs with environment variable
CMAKE_BUILD_PARALLEL_LEVEL [1] on configuration phase. The value
set by that environment variable is used as a number of CPU
threads when it was not specified by user. Number of CPU threads
detected using builtin CMake function [2].

NOTE: CMAKE_BUILD_PARALLEL_LEVEL has been added in a version 3.12.
In earlier CMake versions environment variable is ignored.

1. https://cmake.org/cmake/help/latest/envvar/CMAKE_BUILD_PARALLEL_LEVEL.html
2. https://cmake.org/cmake/help/latest/module/ProcessorCount.html

NO_CHANGELOG=build
NO_DOC=build
NO_TEST=build